Added multiple files for classes, started on grapics-engine
Added files for the classes: flying-unit, rectangle and world, and made
airplane a subclass of flying-unit. which is now a subclass to
rectangle(should be changed later).

Also started on the graphics engine. Now, a frame is defined with a
canvas inside it, and inside the canvas all flying units in *world* are
drawn. There is however a problem with the rotation of the coordinate
system, so this needs to be looked at.