Big Collision update, started collision handling, added some misc...
Did a big expansion of collision. Now, all collisions are able to be
found with a single procedure, and all collisions are now handled with
removing units, buffing them, etc...

This is done through running ($Collision), which does everything needed
during 1 run of the physics engine.

Note however that the procedures initiated by this aren't fully defined
yet, so for example (send plane $Buff 'flying_speed) or (send entity
$Kill) doesn't currently do anything.

The code also hasn't been tested yet, and will be done after the
graphics engine can be run with keypresses.