Added a buff procedure for airplanes, started on respawning
Added a procedure ($Buff buff_type) inside airplane% that buffs the
plane accordingly.

To help this i added the variables base_speed and base_shooting_speed,
so that the original value can be remembered while the others are
altered.

Also added the variable lives, which counts how many lives the player
have left. Also extened ($Kill) For airplanes so that one life is
removed and a timer starts that when finished allows the player to
respawn again.