Added random entity spawner, did some cleanup
Added a random entity spawner. If you call
($Entity_Spawner) an entity will spawn at a "random" location and start
a timer, and a new the procedure won't spawn one again unless until the
timer has run out.

Also did some cleanup in the code, and fixed the bug that happened when
you rotated clockwise and landed on the direction 32.