Added forced respawn, started on winning.
Winning can now be checked to se if someone has won.

If you run ($Victory), the game checks if 2 or more players has started
playing. If they have, it then checks:

1 players with more than 0 lives: that player won!
0 players with more than 0 lives: it's a tie!
>2 players with more than 0 lives: keep playing!

Also added forced respawn, so a player that has started once can't camp
inside the base for to long.